requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
playwright==1.44.0
apscheduler==3.10.4
python-dotenv==1.0.1
pyyaml==6.0.1
rich==13.7.1
fastapi==0.111.0
uvicorn==0.30.1
